Add Script To Geocode Listing Addresses
Search for listing with public addresses that have no or default
latitude & longitudes. Geocode & save either the map or contact address
for these community.

CLI flags exist to remove any coordinates for listings with private
addresses and to remove any default coordinates if we could not geocode
one of the addresses.