Add Script to Export Name & Email of Group Members
The script queries the users of the passed group name or ID, exporting
the best name & the email of each user to a CSV file.

Add the Group & UserGroup tables to the Schema, as well as functions for
pulling a group's users & for building the best name for a user.